add product variant fragment
