+++ Core/GraphQL/API/sell_fragment_productVariant.graphql.swift
@@ -1,0 +1,144 @@
+// @generated
+//  This file was automatically generated and should not be edited.
+
+import Apollo
+import Foundation
+
+/// ApolloType namespace
+public extension ApolloType {
+  struct FragmentProductVariant: GraphQLFragment {
+    /// The raw GraphQL definition of this fragment.
+    public static let fragmentDefinition: String =
+      """
+      fragment FragmentProductVariant on ProductVariant {
+        __typename
+        id
+        externalId
+        price
+        externalId
+        translations {
+          __typename
+          title
+          language
+        }
+      }
+      """
+
+    public static let possibleTypes: [String] = ["ProductVariant"]
+
+    public static var selections: [GraphQLSelection] {
+      return [
+        GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
+        GraphQLField("id", type: .nonNull(.scalar(GraphQLID.self))),
+        GraphQLField("externalId", type: .scalar(String.self)),
+        GraphQLField("price", type: .scalar(Int.self)),
+        GraphQLField("externalId", type: .scalar(String.self)),
+        GraphQLField("translations", type: .list(.object(Translation.selections))),
+      ]
+    }
+
+    public private(set) var resultMap: ResultMap
+
+    public init(unsafeResultMap: ResultMap) {
+      self.resultMap = unsafeResultMap
+    }
+
+    public init(id: GraphQLID, externalId: String? = nil, price: Int? = nil, translations: [Translation?]? = nil) {
+      self.init(unsafeResultMap: ["__typename": "ProductVariant", "id": id, "externalId": externalId, "price": price, "translations": translations.flatMap { (value: [Translation?]) -> [ResultMap?] in value.map { (value: Translation?) -> ResultMap? in value.flatMap { (value: Translation) -> ResultMap in value.resultMap } } }])
+    }
+
+    public var __typename: String {
+      get {
+        return resultMap["__typename"]! as! String
+      }
+      set {
+        resultMap.updateValue(newValue, forKey: "__typename")
+      }
+    }
+
+    public var id: GraphQLID {
+      get {
+        return resultMap["id"]! as! GraphQLID
+      }
+      set {
+        resultMap.updateValue(newValue, forKey: "id")
+      }
+    }
+
+    public var externalId: String? {
+      get {
+        return resultMap["externalId"] as? String
+      }
+      set {
+        resultMap.updateValue(newValue, forKey: "externalId")
+      }
+    }
+
+    public var price: Int? {
+      get {
+        return resultMap["price"] as? Int
+      }
+      set {
+        resultMap.updateValue(newValue, forKey: "price")
+      }
+    }
+
+    public var translations: [Translation?]? {
+      get {
+        return (resultMap["translations"] as? [ResultMap?]).flatMap { (value: [ResultMap?]) -> [Translation?] in value.map { (value: ResultMap?) -> Translation? in value.flatMap { (value: ResultMap) -> Translation in Translation(unsafeResultMap: value) } } }
+      }
+      set {
+        resultMap.updateValue(newValue.flatMap { (value: [Translation?]) -> [ResultMap?] in value.map { (value: Translation?) -> ResultMap? in value.flatMap { (value: Translation) -> ResultMap in value.resultMap } } }, forKey: "translations")
+      }
+    }
+
+    public struct Translation: GraphQLSelectionSet {
+      public static let possibleTypes: [String] = ["ProductVariantTranslation"]
+
+      public static var selections: [GraphQLSelection] {
+        return [
+          GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
+          GraphQLField("title", type: .scalar(String.self)),
+          GraphQLField("language", type: .scalar(Language.self)),
+        ]
+      }
+
+      public private(set) var resultMap: ResultMap
+
+      public init(unsafeResultMap: ResultMap) {
+        self.resultMap = unsafeResultMap
+      }
+
+      public init(title: String? = nil, language: Language? = nil) {
+        self.init(unsafeResultMap: ["__typename": "ProductVariantTranslation", "title": title, "language": language])
+      }
+
+      public var __typename: String {
+        get {
+          return resultMap["__typename"]! as! String
+        }
+        set {
+          resultMap.updateValue(newValue, forKey: "__typename")
+        }
+      }
+
+      public var title: String? {
+        get {
+          return resultMap["title"] as? String
+        }
+        set {
+          resultMap.updateValue(newValue, forKey: "title")
+        }
+      }
+
+      public var language: Language? {
+        get {
+          return resultMap["language"] as? Language
+        }
+        set {
+          resultMap.updateValue(newValue, forKey: "language")
+        }
+      }
+    }
+  }
+}
