+++ Core/GraphQL/GraphQLFiles/sell_fragment_productVariant.graphql
@@ -1,0 +1,10 @@
+fragment FragmentProductVariant on ProductVariant {
+  id
+  externalId
+  price
+  externalId
+  translations {
+    title
+    language
+  }
+}
